refactor: move CursorMove to model/shared, re-export from update::action
