+++ src/app/model/browse/cell_edit.rs
@@ -39,7 +39,7 @@ impl CellEditState {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::update::action::CursorMove;
+    use crate::model::shared::cursor::CursorMove;
 
     #[test]
     fn begin_with_value_sets_active_state_with_copied_values() {
