+++ src/app/model/shared/mod.rs
@@ -1,5 +1,6 @@
 pub mod async_run;
 pub mod confirm_dialog;
+pub mod cursor;
 pub mod db_capabilities;
 pub mod flash_timer;
 pub mod focused_pane;
